core/adobe/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
build==0.9.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.7
cryptography==3.4.6
dataclasses-json==0.6.7
et-xmlfile==1.1.0
filelock==3.15.4
flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
huggingface-hub==0.23.4
idna==3.7
importlib-metadata==8.0.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-postgres==0.0.9
langchain-text-splitters==0.2.2
langsmith==0.1.82
MarkupSafe==2.1.5
marshmallow==3.21.3
mpmath==1.3.0
multidict==6.0.5
multipart==0.2.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.5
orjson==3.10.5
packaging==21.3
pandas==2.2.2
pdfservices-sdk==2.3.1
pep517==0.13.0
pgvector==0.2.5
pillow==10.4.0
polling==0.3.2
polling2==0.5.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
pycparser==2.21
pydantic==2.7.4
pydantic-core==2.18.4
Pygments==2.14.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0
regex==2024.5.15
requests==2.27.1
requests-toolbelt==0.10.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.1
six==1.16.0
SQLAlchemy==2.0.31
sympy==1.12.1
tabulate==0.9.0
tenacity==8.4.2
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.3
triton==2.3.1
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
urllib3==1.26.13
werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2